Refactored class file names and import statements
